+++ lib.commonjs/contract/contract.d.ts
@@ -20,7 +20,7 @@ export declare class BaseContract implements Addressable, EventEmitterable<Contr
      *  The target to connect to.
      *
      *  This can be an address, ENS name or any [[Addressable]], such as
-     *  another contract. To get the resovled address, use the ``getAddress``
+     *  another contract. To get the resolved address, use the ``getAddress``
      *  method.
      */
     readonly target: string | Addressable;
@@ -87,13 +87,13 @@ export declare class BaseContract implements Addressable, EventEmitterable<Contr
     /**
      *  Return the function for a given name. This is useful when a contract
      *  method name conflicts with a JavaScript name such as ``prototype`` or
-     *  when using a Contract programatically.
+     *  when using a Contract programmatically.
      */
     getFunction<T extends ContractMethod = ContractMethod>(key: string | FunctionFragment): T;
     /**
      *  Return the event for a given name. This is useful when a contract
      *  event name conflicts with a JavaScript name such as ``prototype`` or
-     *  when using a Contract programatically.
+     *  when using a Contract programmatically.
      */
     getEvent(key: string | EventFragment): ContractEvent;
     /**