+++ lib.commonjs/crypto/signature.js
@@ -10,6 +10,9 @@ const BN_2 = BigInt(2);
 const BN_27 = BigInt(27);
 const BN_28 = BigInt(28);
 const BN_35 = BigInt(35);
+const BN_N = BigInt("0xfffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141");
+const BN_N_2 = BN_N / BN_2; // Must be integer (floor) division; do NOT shifts
+const inspect = Symbol.for("nodejs.util.inspect.custom");
 const _guard = {};
 function toUint256(value) {
     return (0, index_js_2.zeroPadValue)((0, index_js_2.toBeArray)(value), 32);
@@ -60,7 +63,8 @@ class Signature {
      *  Returns true if the Signature is valid for [[link-eip-2]] signatures.
      */
     isValid() {
-        return (parseInt(this.#s.substring(0, 3)) < 8);
+        const s = BigInt(this.#s);
+        return (s <= BN_N_2);
     }
     /**
      *  The ``v`` value for a signature.
@@ -136,8 +140,26 @@ class Signature {
         this.#v = v;
         this.#networkV = null;
     }
-    [Symbol.for('nodejs.util.inspect.custom')]() {
-        return `Signature { r: "${this.r}", s: "${this._s}"${this.isValid() ? "" : ', valid: "false"'}, yParity: ${this.yParity}, networkV: ${this.networkV} }`;
+    /**
+     *  Returns the canonical signature.
+     *
+     *  This is only necessary when dealing with legacy transaction which
+     *  did not enforce canonical S values (i.e. [[link-eip-2]]. Most
+     *  developers should never require this.
+     */
+    getCanonical() {
+        if (this.isValid()) {
+            return this;
+        }
+        // Compute the canonical signature; s' = N - s, v = !v
+        const s = BN_N - BigInt(this._s);
+        const v = (55 - this.v);
+        const result = new Signature(_guard, this.r, toUint256(s), v);
+        // Populate the networkV if necessary
+        if (this.networkV) {
+            result.#networkV = this.networkV;
+        }
+        return result;
     }
     /**
      *  Returns a new identical [[Signature]].
@@ -159,6 +181,15 @@ class Signature {
             networkV: ((networkV != null) ? networkV.toString() : null),
             r: this.r, s: this._s, v: this.v,
         };
+    }
+    [inspect]() {
+        return this.toString();
+    }
+    toString() {
+        if (this.isValid()) {
+            return `Signature { r: ${this.r}, s: ${this._s}, v: ${this.v} }`;
+        }
+        return `Signature { r: ${this.r}, s: ${this._s}, v: ${this.v}, valid: false }`;
     }
     /**
      *  Compute the chain ID from the ``v`` in a legacy EIP-155 transactions.