+++ lib.commonjs/providers/abstract-signer.d.ts
@@ -42,7 +42,7 @@ export declare abstract class AbstractSigner<P extends null | Provider = null | 
     abstract signTypedData(domain: TypedDataDomain, types: Record<string, Array<TypedDataField>>, value: Record<string, any>): Promise<string>;
 }
 /**
- *  A **VoidSigner** is a class deisgned to allow an address to be used
+ *  A **VoidSigner** is a class designed to allow an address to be used
  *  in any API which accepts a Signer, but for which there are no
  *  credentials available to perform any actual signing.
  *