+++ lib.esm/providers/provider.js
@@ -124,6 +124,9 @@ export function copyRequest(req) {
     if ("kzg" in req) {
         result.kzg = req.kzg;
     }
+    if ("blobWrapperVersion" in req) {
+        result.blobWrapperVersion = req.blobWrapperVersion;
+    }
     if ("blobs" in req && req.blobs) {
         result.blobs = req.blobs.map((b) => {
             if (isBytesLike(b)) {